split into another file
