+++ src/Layer/PaintNodePrivate.hpp
@@ -23,6 +23,7 @@
 #include "Layer/Effects.hpp"
 #include "Layer/Painter.hpp"
 #include "Layer/VSkia.hpp"
+#include "Layer/DisplayList.hpp"
 #include "Utility/HelperMacro.hpp"
 #include "Layer/Core/PaintNode.hpp"
 #include "Renderer.hpp"
@@ -46,90 +47,6 @@
 #include <effects/SkShaderMaskFilter.h>
 #include <effects/SkBlurMaskFilter.h>
 #include <src/shaders/SkPictureShader.h>
-
-class DisplayList
-{
-public:
-  SkPicture* picture() const
-  {
-    return m_displayList ? m_displayList->picture().get() : nullptr;
-  }
-  void playback(Renderer* renderer)
-  {
-    renderer->canvas()->drawPicture(m_displayList->picture());
-  }
-
-  sk_sp<SkShader> asShader() const
-  {
-    return m_displayList;
-  }
-
-  sk_sp<SkImageFilter> asImageFilter()
-  {
-    if (!m_imageFilter)
-    {
-      m_imageFilter = SkImageFilters::Shader(m_displayList);
-    }
-    return m_imageFilter;
-  }
-
-private:
-  friend class DisplayListRecorder;
-  DisplayList(sk_sp<SkPictureShader> shader)
-    : m_displayList(std::move(shader))
-  {
-    ASSERT(m_displayList);
-  }
-  sk_sp<SkPictureShader> m_displayList;
-  sk_sp<SkImageFilter>   m_imageFilter;
-};
-
-class DisplayListRecorder
-{
-public:
-  DisplayListRecorder() = default;
-  DisplayListRecorder(const DisplayListRecorder&) = delete;
-  DisplayListRecorder& operator=(const DisplayListRecorder&) = delete;
-  Renderer*            beginRecording(const SkRect& b, const SkMatrix& matrix)
-  {
-    m_bound = b;
-    m_matrix = matrix;
-    if (!m_rec)
-    {
-      m_rec = std::make_unique<SkPictureRecorder>();
-      auto rt = SkRTreeFactory();
-      auto canvas = m_rec->beginRecording(b, &rt);
-      if (!m_renderer)
-      {
-        m_renderer = std::make_unique<Renderer>();
-      }
-      m_renderer->setCanvas(canvas);
-    }
-    return renderer();
-  }
-
-  Renderer* renderer() const
-  {
-    return m_renderer.get();
-  }
-  DisplayList finishRecording()
-  {
-    auto maskShader = sk_make_sp<SkPictureShader>(
-      m_rec->finishRecordingAsPicture(),
-      SkTileMode::kClamp,
-      SkTileMode::kClamp,
-      SkFilterMode::kNearest,
-      &m_bound);
-    ASSERT(maskShader);
-    return DisplayList(maskShader);
-  }
-
-private:
-  SkMatrix                           m_matrix;
-  SkRect                             m_bound;
-  std::unique_ptr<Renderer>          m_renderer;
-  std::unique_ptr<SkPictureRecorder> m_rec;
-};
 
 namespace VGG::layer
 {
